+++ Samples/BlazorExample/BlazorExample/BlazorExample.Client/Program.cs
@@ -10,7 +10,7 @@ builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.
 
 builder.Services.AddCsla(o => o
   .AddBlazorWebAssembly(o => o.SyncContextWithServer = true)
-  .Security(o => o.FlowSecurityPrincipalFromClient = true)
+  .Security(o => o.FlowSecurityPrincipalFromClient = false)
   .DataPortal(o => o.ClientSideDataPortal(o => o
     .UseHttpProxy(o => o.DataPortalUrl = "/api/DataPortal"))));